+++ src/functions/markerClusterGroup.ts
@@ -1,13 +1,10 @@
 import type { Ref, SetupContext, ExtractPropTypes } from 'vue'
 import type { MarkerClusterGroup, LeafletEventHandlerFnMap, Layer } from 'leaflet'
 import type { ILayerDefinition } from '@vue-leaflet/vue-leaflet/dist/src/types/interfaces/ILayerDefinition.d.ts'
-
+import type { LeafletEventKeys } from '@/types/eventKeys'
 import { provide } from 'vue'
 import { Functions, Utilities, InjectionKeys } from '@vue-leaflet/vue-leaflet'
 import { debounce } from './utils'
-
-type MissingEventHandlerKeys = 'clustermouseover' | 'clustermouseout' | 'clusterclick'
-export type LeafletEventKeys = Array<keyof LeafletEventHandlerFnMap | MissingEventHandlerKeys>
 
 const { featureGroupProps, setupFeatureGroup } = Functions.FeatureGroup
 const { AddLayerInjection, RemoveLayerInjection } = InjectionKeys