+++ src/types/eventKeys.ts
@@ -1,0 +1,3 @@
+import type { LeafletEventHandlerFnMap } from 'leaflet'
+
+export type LeafletEventKeys = Array<keyof LeafletEventHandlerFnMap>