+++ src/functions/markerClusterGroup.ts
@@ -49,7 +49,7 @@ export const markerClusterGroupProps = {
    * current map zoom and returns the maximum cluster radius in pixels.
    **/
   maxClusterRadius: {
-    type: Number,
+    type: [Number, Function],
     default: 80
   },
   /**